Added Grid Style Project
I couldn't get the other project to run,  I used a template and some
code I created for a Youtube app.  This will render the RSS Feeds and
Give us a Details Page for Each.  We can change it to give us categories
if we need in the future.  We still need to map the show name to the
IMDB or EZTV to get the description and some images for the Show.